Major rewrite
Added a rayon feature, and hidden all the parallel calls behind it.
Added a binary version of the library, that if run, would run the big_matrix2 calculation some times, not saving the result.
This is to use in profiling tools.

Instead of copying/clonning the columns each time the ici_kendall_tau is called, they are passed as read-only views instead.

Currently, the output is wrong..